uint -> uint256

+++ contracts/ExamResultProcessor.sol
@@ -2,31 +2,29 @@
 pragma solidity 0.8.16;
 
 contract ExamResultProcessor {
-
     struct ResultData {
-        uint studentId;
+        uint256 studentId;
         string subject;
-        uint score;
+        uint256 score;
     }
-    uint constant passMark = 60;
+    uint256 constant passMark = 60;
 
     ResultData currentResult;
-    mapping (uint => mapping (string => bool)) public subjectsPassed;
-    mapping (uint => uint) public totalScore;
+    mapping(uint256 => mapping(string => bool)) public subjectsPassed;
+    mapping(uint256 => uint256) public totalScore;
 
     function process(ResultData[] memory results) public {
-        for (uint i=0; i<results.length; i++) {
+        for (uint256 i = 0; i < results.length; i++) {
             currentResult = results[i];
             processCurrentResult();
         }
     }
 
     function processCurrentResult() private {
-        uint studentId = currentResult.studentId;
+        uint256 studentId = currentResult.studentId;
         string memory subject = currentResult.subject;
-        uint score = currentResult.score;
+        uint256 score = currentResult.score;
         subjectsPassed[studentId][subject] = score >= passMark;
         totalScore[studentId] += score;
     }
-
 }
